requirements.txt
Flask
jinja2
flask-socketio
eventlet
python-statemachine
